DIV-5613 Allow DN Drafted state link (#370)
