+++ config/default.yml
@@ -141,6 +141,7 @@ coRespRespondableStates: [
   'DefendedDivorce',
   'AwaitingDecreeNisi',
   'DNAwaiting',
+  'DNDrafted',
   'AwaitingLegalAdvisorReferral'
 ]
 
@@ -161,6 +162,7 @@ caseStates:
   DefendedDivorce: 'DefendedDivorce'
   DivorceGranted: 'DivorceGranted'
   AosAwaitingSol: 'AosAwaitingSol'
+  DNDrafted: 'DNDrafted'
 
 features:
   idam: false
